+++ packages/fetch-proxy/src/lib/fetch-proxy.test.ts
@@ -7,6 +7,7 @@ async function runProxy(
   request: Request,
   target: string | URL,
   options?: FetchProxyOptions,
+  init?: RequestInit,
 ): Promise<{ request: Request; response: Response }> {
   let outgoingRequest: Request;
   let proxy = createFetchProxy(target, {
@@ -17,7 +18,7 @@ async function runProxy(
     },
   });
 
-  let proxyResponse = await proxy(request);
+  let proxyResponse = await proxy(request, init);
 
   assert.ok(outgoingRequest!);
 
@@ -144,5 +145,43 @@ describe('fetch proxy', () => {
     assert.equal(setCookie.length, 2);
     assert.equal(setCookie[0], 'name=value; Domain=remix.run:3000; Path=/rsc/search');
     assert.equal(setCookie[1], 'name2=value2; Domain=remix.run:3000; Path=/rsc');
+  });
+
+  it('deletes content-encoding and content-length headers', async () => {
+    let { response } = await runProxy(
+      new Request('http://shopify.com/search?q=remix'),
+      'https://remix.run:3000/rsc',
+      {
+        async fetch() {
+          return new Response(null, {
+            headers: [
+              ['Content-Encoding', 'gzip'],
+              ['Content-Length', '1337'],
+            ],
+          });
+        },
+      },
+    );
+
+    assert.equal(response.headers.has('Content-Encoding'), false);
+    assert.equal(response.headers.has('Content-Length'), false);
+  });
+
+  it('applies RequestInit properties to the proxied request', async () => {
+    let { request } = await runProxy(
+      new Request('http://shopify.com/search?q=remix', {
+        redirect: 'manual',
+        referrer: 'http://example.com/referer',
+      }),
+      'https://remix.run:3000/rsc',
+      undefined,
+      {
+        redirect: 'manual',
+        referrer: 'http://example.com/referer',
+      },
+    );
+
+    assert.equal(request.redirect, 'manual');
+    assert.equal(request.referrer, 'http://example.com/referer');
   });
 });